<SWCORE-12188> Creating a MultiReader with SampleType::Undefined as value read type will infer it from the descriptor

+++ examples/cpp/quick_start/quick_start_full.cpp
@@ -189,7 +189,7 @@ int streamReaderTest(int /*argc*/, const char* /*argv*/ [])
         now = std::chrono::system_clock::now();
     }
 
-    for (int i = 0; i < std::size(counts); ++i)
+    for (int i = 0; i < static_cast<int>(std::size(counts)); ++i)
     {
         if (counts[i] != 0)
         {
